requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-bootstrap3==21.2
django-crispy-forms==1.14.0
django-filter==21.1
django-tables2==2.4.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
whitenoise==5.3.0
